Added validation to form to ensure all fields have been entered
insert.php now contains one large if statement with multiple || to check all
required fields have a value. Added code to intake.php that when an error on
submission occurs, shows the error message, and retains all previous entries
so the user does not have to fill the form out entirely again.