Added a way to deal with successful form submissions, fixed functions error
Found a way that if the form was successfully added to the database, the
page is redirected back to the intake page but only displays the header
along with a success message.

Also fixed an error in the functions page. Functions that make a connection
to the database were including the wrong file, causing the functions to fail.
These functions have been corrected and now include connection.php.