Refactored code to use a router
Added two new classes.
The Router class processes a router file that is just an array of routes.
It builds up the array inside the class and when the direct method is called,
the array searched against the given uri.

The Request class proccess a uri request and trims the input, and uses that
input to search through the Router class direct method.

Added a views folder which houses the html views that the user will see.

Moved header and footer partials out of the inc folder into a partials
folder located inside the views folder.

Added a controllers folder. The controllers currently just return the
view assosciated with the controller. In the future the controllers will
do more to interact with the data from the database.