Refactored database connection
Added two classes to handle the database connection and another to build
up the query sent to the database.
Added a bootstrap file to handle all the required files, as well as making
the initial connection to the database.
Added a config file to handle sensitive information relating to the database.
Currently the site is able to add an intake to the database successfully.
Known issues:
*Date format entered into the database is currently a string and not
date format
**Search page is broken since a search method has not been added to the
QueryBuilder class
***Need to refactor the bootstrap file to take advantage of autoloading