Refactored database
Added two classes.
1. Connection is responsible for establishing a connection to the database.
2. QueryBuilder is responsible for building up the query statments passed
to the database. QueryBuilder* currently only adds an intake to the database.

Added a config file that contains settings for connecting to the database.
Added a bootstrap file** that brings in all of the required files and
passes them on to the index. Also makes the connection to the database.

Known Issues/Refactors:
* Search page is currently broken since there is only a method defined in
QueryBuilder to insert into the database. Need to add a search method.
** Need to implement autoloading feature.