Refactored to use a basic dependency injector
The DI is handled in the new App class. App currently accepts the config
file and stores it, as well as the database connection. When either of
these files need to be accessed, they can be called using App::get().

This cleans up the bootstrap file some and adds some flexibility when
adding to the DI and when calling form it.